+++ R/5.1-fs.R
@@ -14,26 +14,38 @@
 #'  Set \code{top = 0} to include all features. A numeric vector can also be used
 #'  to indicate specific features by location, similar to a character vector.
 #' @param uniqueFx A function call unique to the method.
+#' @param keep A numeric scalar. Specifies the number of top features that should get
+#'  returned by the feature selection method. Use of \code{keep} is generally not
+#'  recommended, but can speed up analyses of large data.
 #' @param ... Arguments passed to the detailed function.
 #' @return Returns an \code{ExprsArray} object.
 #' @export
-fs. <- function(object, top, uniqueFx, ...){
+fs. <- function(object, top, uniqueFx, keep, ...){
 
   # Convert top input to explicit feature reference
   if(class(top) == "numeric"){
     if(length(top) == 1){
       if(top > nrow(object@exprs)) top <- 0
-      if(top == 0) top <- nrow(object@exprs)
-      top <- rownames(object@exprs[1:top, ])
+      if(top == 0){
+        top <- rownames(object@exprs) # keep for uniqueFx
+        data <- t(object@exprs)
+      }else{
+        top <- rownames(object@exprs)[1:top]
+        data <- t(object@exprs[top, , drop = FALSE])
+      }
     }else{
-      top <- rownames(object@exprs[top, ])
+      top <- rownames(object@exprs)[top] # when top is numeric vector
+      data <- t(object@exprs[top, , drop = FALSE])
     }
+  }else{
+    # top <- top # feature names already specified
+    data <- t(object@exprs[top, , drop = FALSE])
   }
 
   # Build data from top subset for uniqueFx
-  data <- t(object@exprs[top, ])
   outcome <- object@annot$defineCase
   final <- do.call("uniqueFx", list(data, outcome, top, ...))
+  if(keep != 0) final <- final[1:keep]
 
   # Append uniqueFx results to object
   if(class(final) == "character"){ # fill @preFilter slot
@@ -56,7 +68,7 @@ fs. <- function(object, top, uniqueFx, ...){
 #' @inheritParams fs.
 #' @return Returns an \code{ExprsArray} object.
 #' @export
-fsSample <- function(object, top = 0, ...){ # args to sample
+fsSample <- function(object, top = 0, keep = 0, ...){ # args to sample
 
   classCheck(object, "ExprsArray",
              "This function is applied to the results of ?exprso.")
@@ -65,7 +77,7 @@ fsSample <- function(object, top = 0, ...){ # args to sample
       uniqueFx = function(data, outcome, top, ...){
 
         sample(top, ...)
-      }, ...)
+      }, keep, ...)
 }
 
 #' Null Feature Selection
@@ -75,7 +87,7 @@ fsSample <- function(object, top = 0, ...){ # args to sample
 #' @inheritParams fs.
 #' @return Returns an \code{ExprsArray} object.
 #' @export
-fsNULL <- function(object, top = 0, ...){ # args to NULL
+fsNULL <- function(object, top = 0, keep = 0, ...){ # args to NULL
 
   classCheck(object, "ExprsArray",
              "This function is applied to the results of ?exprso.")
@@ -84,7 +96,7 @@ fsNULL <- function(object, top = 0, ...){ # args to NULL
       uniqueFx = function(data, outcome, top, ...){
 
         top
-      }, ...)
+      }, keep, ...)
 }
 
 #' Select Features by Explicit Reference
@@ -97,7 +109,7 @@ fsNULL <- function(object, top = 0, ...){ # args to NULL
 #'  This preserves the feature order otherwise. Argument for \code{fsInclude} only.
 #' @return Returns an \code{ExprsArray} object.
 #' @export
-fsInclude <- function(object, top = 0, include){
+fsInclude <- function(object, top = 0, keep = 0, include){
 
   classCheck(object, "ExprsArray",
              "This function is applied to the results of ?exprso.")
@@ -115,7 +127,7 @@ fsInclude <- function(object, top = 0, include){
 
         index <- colnames(data) %in% include
         c(include, colnames(data)[!index])
-      }, include)
+      }, keep, include)
 }
 
 #' Select Features by ANOVA
@@ -126,7 +138,7 @@ fsInclude <- function(object, top = 0, include){
 #' @inheritParams fs.
 #' @return Returns an \code{ExprsArray} object.
 #' @export
-fsANOVA <- function(object, top = 0, ...){ # args to aov
+fsANOVA <- function(object, top = 0, keep = 0, ...){ # args to aov
 
   classCheck(object, c("ExprsBinary", "ExprsMulti"),
              "This feature selection method only works for classification tasks.")
@@ -145,20 +157,21 @@ fsANOVA <- function(object, top = 0, ...){ # args to aov
         }
 
         top[order(p)]
-      }, ...)
+      }, keep, ...)
 }
 
 #' Select Features by Statistical Testing
 #'
-#' \code{fsStats} selects features using the \code{t.test} or \code{ks.test}
+#' \code{fsStats} selects features using a base R statistics
 #'  function (toggled by the \code{how} argument).
 #'
 #' @inheritParams fs.
-#' @param how A character string. Toggles between the "t.test" and
-#'  "ks.test" method. Argument for \code{fsStats} only.
+#' @param how A character string. Toggles between the "t.test", "ks.test",
+#'  "wilcox.test", and "var.test" methods.
 #' @return Returns an \code{ExprsArray} object.
 #' @export
-fsStats <- function(object, top = 0, how = c("t.test", "ks.test"), ...){ # args to t.test or ks.test
+fsStats <- function(object, top = 0, keep = 0,
+                    how = c("t.test", "ks.test", "wilcox.test", "var.test"), ...){ # args to base R function
 
   classCheck(object, "ExprsBinary",
              "This feature selection method only works for binary classification tasks.")
@@ -180,7 +193,7 @@ fsStats <- function(object, top = 0, how = c("t.test", "ks.test"), ...){ # args 
             })
           }
           top[order(p)]
-        }, ...)
+        }, keep, ...)
 
   }else if(how[1] == "ks.test"){
 
@@ -199,7 +212,45 @@ fsStats <- function(object, top = 0, how = c("t.test", "ks.test"), ...){ # args 
             })
           }
           top[order(p)]
-        }, ...)
+        }, keep, ...)
+
+  }else if(how[1] == "wilcox.test"){
+
+    fs.(object, top,
+        uniqueFx = function(data, outcome, top, ...){
+
+          dca <- data[outcome == "Case", ]
+          dco <- data[outcome == "Control", ]
+          p <- vector("numeric", length(top))
+          for(i in 1:ncol(data)){
+            tryCatch({
+              p[i] <- wilcox.test(dca[, i], dco[, i], ...)$p.value
+            }, error = function(e){
+              cat("fsStats failed for feature: ", top[i], ". Setting p(x)=1...\n")
+              p[i] <- 1
+            })
+          }
+          top[order(p)]
+        }, keep, ...)
+
+  }else if(how[1] == "var.test"){
+
+    fs.(object, top,
+        uniqueFx = function(data, outcome, top, ...){
+
+          dca <- data[outcome == "Case", ]
+          dco <- data[outcome == "Control", ]
+          p <- vector("numeric", length(top))
+          for(i in 1:ncol(data)){
+            tryCatch({
+              p[i] <- var.test(dca[, i], dco[, i], ...)$p.value
+            }, error = function(e){
+              cat("fsStats failed for feature: ", top[i], ". Setting p(x)=1...\n")
+              p[i] <- 1
+            })
+          }
+          top[order(p)]
+        }, keep, ...)
 
   }else{
 
@@ -215,7 +266,7 @@ fsStats <- function(object, top = 0, how = c("t.test", "ks.test"), ...){ # args 
 #' @inheritParams fs.
 #' @return Returns an \code{ExprsArray} object.
 #' @export
-fsCor <- function(object, top = 0, ...){ # args to cor
+fsCor <- function(object, top = 0, keep = 0, ...){ # args to cor
 
   classCheck(object, "RegrsArray",
              "This feature selection method only works for continuous outcome tasks.")
@@ -228,7 +279,7 @@ fsCor <- function(object, top = 0, ...){ # args to cor
           r[i] <- cor(data[,i], outcome)
         }
         top[rev(order(r))]
-      }, ...)
+      }, keep, ...)
 }
 
 #' Reduce Dimensions by PCA
@@ -240,7 +291,7 @@ fsCor <- function(object, top = 0, ...){ # args to cor
 #' @inheritParams fs.
 #' @return Returns an \code{ExprsArray} object.
 #' @export
-fsPrcomp <- function(object, top = 0, ...){ # args to prcomp
+fsPrcomp <- function(object, top = 0, keep = 0, ...){ # args to prcomp
 
   classCheck(object, "ExprsArray",
              "This function is applied to the results of ?exprso.")
@@ -257,7 +308,7 @@ fsPrcomp <- function(object, top = 0, ...){ # args to prcomp
         #  when calling modHistory
         list(t(reductionModel$x),
              reductionModel)
-      }, ...)
+      }, keep, ...)
 }
 
 #' Select Features by Recursive Feature Elimination
@@ -268,7 +319,7 @@ fsPrcomp <- function(object, top = 0, ...){ # args to prcomp
 #' @inheritParams fs.
 #' @return Returns an \code{ExprsArray} object.
 #' @export
-fsPathClassRFE <- function(object, top = 0, ...){ # args to fit.rfe
+fsPathClassRFE <- function(object, top = 0, keep = 0, ...){ # args to fit.rfe
 
   packageCheck("pathClass")
   classCheck(object, "ExprsBinary",
@@ -290,7 +341,7 @@ fsPathClassRFE <- function(object, top = 0, ...){ # args to fit.rfe
         final <- merge(data.frame("new" = rfe$features), key, sort = FALSE)$old
         if(length(final) < 2) stop("Uh oh! fsPathClassRFE did not find enough features!")
         as.character(final)
-      }, ...)
+      }, keep, ...)
 }
 
 #' Select Features by Moderated t-test
@@ -302,7 +353,7 @@ fsPathClassRFE <- function(object, top = 0, ...){ # args to fit.rfe
 #' @inheritParams fs.
 #' @return Returns an \code{ExprsArray} object.
 #' @export
-fsEbayes <- function(object, top = 0, ...){ # args to ebayes
+fsEbayes <- function(object, top = 0, keep = 0, ...){ # args to ebayes
 
   packageCheck("limma")
   classCheck(object, c("ExprsBinary", "ExprsMulti"),
@@ -316,7 +367,7 @@ fsEbayes <- function(object, top = 0, ...){ # args to ebayes
         ebaye <- limma::eBayes(fit)
         tt <- limma::topTableF(ebaye, number = ncol(data))
         rownames(tt)
-      }, ...)
+      }, keep, ...)
 }
 
 #' Selects Features by Exact Test
@@ -333,7 +384,7 @@ fsEbayes <- function(object, top = 0, ...){ # args to ebayes
 #' @inheritParams fs.
 #' @return Returns an \code{ExprsArray} object.
 #' @export
-fsEdger <- function(object, top = 0, ...){ # args to exactTest
+fsEdger <- function(object, top = 0, keep = 0, ...){ # args to exactTest
 
   packageCheck("edgeR")
   classCheck(object, "ExprsBinary",
@@ -348,7 +399,7 @@ fsEdger <- function(object, top = 0, ...){ # args to exactTest
         et <- edgeR::exactTest(y)
         tt <- as.data.frame(edgeR::topTags(et, n = nrow(et)))
         rownames(tt)
-      }, ...)
+      }, keep, ...)
 }
 
 #' Select Features by mRMR
@@ -362,7 +413,7 @@ fsEdger <- function(object, top = 0, ...){ # args to exactTest
 #' @inheritParams fs.
 #' @return Returns an \code{ExprsArray} object.
 #' @export
-fsMrmre <- function(object, top = 0, ...){ # args to mRMR.classic
+fsMrmre <- function(object, top = 0, keep = 0, ...){ # args to mRMR.classic
 
   packageCheck("mRMRe")
   classCheck(object, "ExprsBinary",
@@ -392,7 +443,39 @@ fsMrmre <- function(object, top = 0, ...){ # args to mRMR.classic
         # Use "make.names" key to return to original row.names
         final <- merge(data.frame("new" = final), key, sort = FALSE)$old
         as.character(final)
-      }, ...)
+      }, keep, ...)
+}
+
+#' Select Features by Rank Product Analysis
+#'
+#' \code{fsRankProd} selects features using the \code{RankProducts} function
+#'  from the \code{RankProd} package.
+#'
+#' @inheritParams fs.
+#' @return Returns an \code{ExprsArray} object.
+#' @export
+fsRankProd <- function(object, top = 0, keep = 0, ...){ # args to RankProducts
+
+  packageCheck("RankProd")
+  classCheck(object, "ExprsBinary",
+             "This feature selection method only works for binary classification tasks.")
+
+  fs.(object, top,
+      uniqueFx = function(data, outcome, top, ...){
+
+        args <- getArgs(...)
+        args <- defaultArg("logged", FALSE, args)
+
+        cl <- ifelse(outcome == "Control", 0, 1)
+        args <- append(list("data" = t(data), "cl" = cl), args)
+        sink(tempfile())
+        final <- do.call(RankProd::RankProducts, args)
+        sink()
+
+        # Get two-tailed p-value
+        p <- apply(final$pval, 1, min)
+        top[order(p)]
+      }, keep, ...)
 }
 
 #' Select Features by Differential Proportionality Analysis
@@ -401,31 +484,60 @@ fsMrmre <- function(object, top = 0, ...){ # args to mRMR.classic
 #'  from the \code{propr} package.
 #'
 #' @inheritParams fs.
+#' @param modRatios A logical scalar. Toggles whether to compute theta from
+#'  the feature ratios as provided. Set \code{modRatios = TRUE} if data were
+#'  recasted by a prior \code{modRatios} call. If \code{TRUE}, the \code{alpha}
+#'  and \code{weighted} arguments will not work.
 #' @return Returns an \code{ExprsArray} object.
 #' @export
-fsPropd <- function(object, top = 0, ...){ # args to propd
+fsPropd <- function(object, top = 0, keep = 0, modRatios = FALSE, ...){ # args to propd
 
   packageCheck("propr")
   classCheck(object, "ExprsBinary",
              "This feature selection method only works for binary classification tasks.")
 
-  fs.(object, top,
-      uniqueFx = function(data, outcome, top, ...){
+  if(modRatios){
 
-        # Order pairs by theta
-        pd <- propr::propd(data, outcome)
-        pd@theta <- pd@theta[order(pd@theta$theta),]
+    fs.(object, top,
+        uniqueFx = function(data, outcome, top, ...){
 
-        # Index features by when they first appear
-        nrows <- nrow(pd@theta)
-        index <- floor(seq(1, nrows+.5, .5))
-        odds <- as.logical(1:(nrows*2) %% 2)
-        index[odds] <- index[odds] + nrows
-        join <- c(pd@theta$Partner, pd@theta$Pair)
-        join <- join[index]
+          # Set up groups and sizes
+          grp1 <- as.character(outcome) == "Control"
+          p1 <- sum(grp1) - 1
+          grp2 <- !grp1
+          p2 <- sum(grp2) - 1
+          p <- p1 + p2 + 1
 
-        # Rank features by first appearance
-        rankedfeats <- unique(join)
-        top[rankedfeats]
-      }, ...)
+          # Calculate theta
+          thetas <- apply(data, 2, function(x){
+            (p1 * stats::var(x[grp1]) + p2 * stats::var(x[grp2])) / (p * stats::var(x))
+          })
+
+          # Sort results
+          top <- names(thetas[order(thetas)])
+          top
+        }, keep, ...)
+
+  }else{
+
+    fs.(object, top,
+        uniqueFx = function(data, outcome, top, ...){
+
+          # Order pairs by theta
+          pd <- suppressMessages(propr::propd(data, outcome))
+          pd@theta <- pd@theta[order(pd@theta$theta),]
+
+          # Index features by when they first appear
+          nrows <- nrow(pd@theta)
+          index <- floor(seq(1, nrows+.5, .5))
+          odds <- as.logical(1:(nrows*2) %% 2)
+          index[odds] <- index[odds] + nrows
+          join <- c(pd@theta$Partner, pd@theta$Pair)
+          join <- join[index]
+
+          # Rank features by first appearance
+          rankedfeats <- unique(join)
+          top[rankedfeats]
+        }, keep, ...)
+  }
 }