+++ tests/testthat/test-build.R
@@ -70,7 +70,35 @@ test_that("built models can predict correct classes", {
   )
 
   set.seed(12345)
-  mach <- buildANN(array.train, top = 2, size = 3, decay = 1)
+  mach <- buildANN(array.train, top = 2)
+  expect_equal(
+    as.character(predict(mach, array.test)@pred),
+    array.test$defineCase
+  )
+
+  set.seed(12345)
+  mach <- buildDT(array, cp = .1)
+  expect_equal(
+    as.character(predict(mach, array)@pred),
+    array$defineCase
+  )
+
+  set.seed(12345)
+  mach <- buildRF(array.train)
+  expect_equal(
+    as.character(predict(mach, array.test)@pred),
+    array.test$defineCase
+  )
+
+  set.seed(12345)
+  mach <- buildFRB(array.train)
+  expect_equal(
+    as.character(predict(mach, array.test)@pred),
+    array.test$defineCase
+  )
+
+  set.seed(12345)
+  mach <- buildLR(array.train)
   expect_equal(
     as.character(predict(mach, array.test)@pred),
     array.test$defineCase
@@ -104,7 +132,28 @@ test_that("built models can detect wrong classes", {
   )
 
   set.seed(12345)
-  mach <- buildANN(array.train, top = 2, size = 3, decay = 1)
+  mach <- buildANN(array.train, top = 2)
+  expect_equal(
+    calcStats(predict(mach, array.test), aucSkip = TRUE)$acc,
+    .9
+  )
+
+  set.seed(12345)
+  mach <- buildRF(array.train)
+  expect_equal(
+    calcStats(predict(mach, array.test), aucSkip = TRUE)$acc,
+    .9
+  )
+
+  set.seed(12345)
+  mach <- buildFRB(array.train)
+  expect_equal(
+    calcStats(predict(mach, array.test), aucSkip = TRUE)$acc,
+    .9
+  )
+
+  set.seed(12345)
+  mach <- buildLR(array.train)
   expect_equal(
     calcStats(predict(mach, array.test), aucSkip = TRUE)$acc,
     .9